Remove bit vector prelude and update docs
