+++ bench/src/mem_bitvec.rs
@@ -1,6 +1,4 @@
-use jerky::bit_vector::bit_vector::BitVectorBuilder;
-use jerky::bit_vector::rank9sel::inner::Rank9SelIndex;
-use jerky::bit_vector::BitVector;
+use jerky::bit_vector::{BitVector, BitVectorBuilder, Rank9SelIndex};
 use rand::{Rng, SeedableRng};
 use rand_chacha::ChaChaRng;
