+++ bench/src/mem_chrseq.rs
@@ -1,6 +1,4 @@
-use jerky::bit_vector::bit_vector::BitVectorBuilder;
-use jerky::bit_vector::rank9sel::inner::Rank9SelIndex;
-use jerky::bit_vector::{BitVector, NoIndex, Rank};
+use jerky::bit_vector::{BitVector, BitVectorBuilder, NoIndex, Rank, Rank9SelIndex};
 use jerky::char_sequences::WaveletMatrix;
 use jerky::int_vectors::CompactVector;
